fix(core): restrict setAttributes to step body in V5 MVP
The previous design used a 'use step' indirection inside @workflow/core
so setAttributes could be called from both workflow and step bodies via
a single SDK surface. That broke nextjs-webpack Local Dev: the deferred-
entries discoverer in webpack dev mode walks transitive imports from
'use step' files, and putting a step file inside @workflow/core/dist
pulled host-side world adapters and @vercel/queue into the
step-discovery graph. Webpack's regex-based import extractor then blew
the call stack with "RangeError: Maximum call stack size exceeded at
RegExpStringIterator.next" on tarball-installed deployments.
runtime/start.ts and runtime/run.ts get away with the same directive
because they're never reachable from packages/core/src/workflow/index.ts
(the VM bundle entry); ours was.

A host-side bridge comparable to sleep would have fixed it but is
substantial wiring for a feature whose end state (event-sourced
attr_set) replaces the bridge mechanism entirely. Pragmatic MVP path:
restrict to step body and let users wrap in a step explicitly. Full
5.0.0 lifts the restriction via attr_set events through the workflow
controller; SDK signature is stable across the cutover.

- Workflow-VM-side setAttributes throws FatalError with wrap-in-step
  instructions
- Step-side setAttributes works (validates, dispatches, world-detects)
- set-attributes-shared.ts is now pure validation; no 'use step', no
  world imports
- Test updated to assert the FatalError on workflow-body calls
- Changelog MDX updated with the new scope + a "Why workflow-body
  dispatch is deferred" implementation note
