+++ packages/core/src/set-attributes-shared.ts
@@ -4,7 +4,6 @@ import {
   AttributeValidationError,
   validateAttributeChanges,
 } from '@workflow/world';
-import { getWorldLazy } from './runtime/get-world-lazy.js';
 
 // IMPORTANT: this module is imported by both the workflow-VM bundle and
 // the step/host bundle. It must NOT import anything that pulls in
@@ -13,48 +12,18 @@ import { getWorldLazy } from './runtime/get-world-lazy.js';
 // constraint, and an offending transitive import will fail the build
 // with "You are attempting to use 'node:async_hooks' which is a Node.js
 // module. Node.js modules are not available in workflow functions."
-
-let unsupportedWorldWarned = false;
-
-function warnUnsupportedWorldOnce(worldName?: string): void {
-  if (unsupportedWorldWarned) return;
-  unsupportedWorldWarned = true;
-  // biome-ignore lint/suspicious/noConsole: surface in user terminals
-  console.warn(
-    `[workflow] setAttributes: the current world implementation${
-      worldName ? ` (${worldName})` : ''
-    } does not implement experimentalSetAttributes; this call (and any subsequent setAttributes calls in this process) is a no-op. Attributes will become available once the world adapter adds support.`
-  );
-}
-
-/**
- * Internal step that performs the actual `experimentalSetAttributes`
- * call against the World. The `'use step'` directive means the function
- * body runs in a step (Node.js) context regardless of which surface
- * invoked it:
- *
- * - Called from a workflow body: dispatched as a step through the
- *   workflow controller, recorded in the event log
- *   (`step_created`/`step_completed`), replayed deterministically on
- *   resume.
- * - Called from a step body (or from Node.js outside a workflow): runs
- *   inline. Step bodies cannot nest steps; the directive is a no-op
- *   there and the function body executes as a plain async function.
- *
- * @internal
- */
-export async function setAttributesStep(
-  runId: string,
-  changes: AttributeChange[]
-): Promise<void> {
-  'use step';
-  const world = await getWorldLazy();
-  if (typeof world.runs.experimentalSetAttributes !== 'function') {
-    warnUnsupportedWorldOnce((world as any)?.name);
-    return;
-  }
-  await world.runs.experimentalSetAttributes(runId, changes);
-}
+//
+// It must ALSO not contain a `'use step'` directive: the deferred-entry
+// discoverer in `@workflow/next/builder-deferred.ts` walks transitive
+// imports from `'use step'` files, and adding a step file to
+// `@workflow/core/dist/` puts the host-side `world.ts` (and its
+// transitive `@vercel/queue` + adapter imports) inside the workflow
+// step-discovery graph. That triggers a stack overflow inside webpack
+// dev mode's regex-based import extractor on tarball-installed
+// deployments. The MVP works around this by performing the world
+// dispatch from the SDK helper (step body only); workflow-body use is
+// supported only via the host-side bridge in `set-attributes.ts`,
+// which calls into an actual step via `registerStepFunction`.
 
 /**
  * Validate and normalize a `setAttributes(record)` call. Returns the
