+++ packages/core/src/workflow/set-attributes.ts
@@ -1,43 +1,41 @@
-import { NotInWorkflowOrStepContextError } from '../context-violation-error.js';
-import {
-  normalizeSetAttributesInput,
-  setAttributesStep,
-} from '../set-attributes-shared.js';
-import { WORKFLOW_CONTEXT_SYMBOL } from './get-workflow-metadata.js';
+import { FatalError } from '@workflow/errors';
 
 /**
- * Workflow-VM-side `setAttributes`. The exported symbol from
- * `@workflow/core/_workflow` (which the umbrella `workflow` package
- * re-exports under the workflow bundle).
+ * Workflow-VM-side `setAttributes` for V5 MVP.
  *
- * Only reads from `globalThis[WORKFLOW_CONTEXT_SYMBOL]` — the workflow
- * VM bundle is built without `node:async_hooks`, so it must not import
- * `contextStorage`. The step/host-side variant
- * (`packages/core/src/set-attributes.ts`) handles both contexts.
+ * In V5 MVP, `setAttributes` is restricted to **step bodies** — calling
+ * it from inside a workflow body throws `FatalError`. The restriction
+ * exists because dispatching the call through the workflow controller
+ * requires either a `'use step'`-tagged helper inside `@workflow/core`
+ * (which trips the deferred-entry discoverer in `nextjs-webpack` dev
+ * mode by adding host-side world adapters to the step-discovery graph)
+ * or host-side bridge wiring comparable in scope to `sleep`. Neither is
+ * worth the complexity for the MVP, given that wrapping a single line
+ * in a `'use step'` function in user code is trivial:
  *
- * The dispatch step (`setAttributesStep`) carries the `'use step'`
- * directive; the SWC plugin rewrites the call into a workflow
- * controller request so the body executes in the step (Node) context.
+ * ```ts
+ * async function setAttrs(attrs: Record<string, string | undefined>) {
+ *   'use step';
+ *   await setAttributes(attrs);
+ * }
  *
- * See the `attributes-mvp` changelog entry for the full API spec.
+ * export async function myWorkflow() {
+ *   'use workflow';
+ *   await setAttrs({ phase: 'init' });
+ * }
+ * ```
+ *
+ * The full Workflow Attributes feature in 5.0.0 dispatches via
+ * `attr_set` events through the workflow controller and lifts this
+ * restriction. The SDK signature does not change.
  */
 export async function setAttributes(
-  attrs: Record<string, string | undefined>
+  _attrs: Record<string, string | undefined>
 ): Promise<void> {
-  const changes = normalizeSetAttributesInput(attrs);
-  if (!changes) return;
-
-  const workflowCtx = (globalThis as any)[WORKFLOW_CONTEXT_SYMBOL] as
-    | { workflowRunId?: string }
-    | undefined;
-  const runId = workflowCtx?.workflowRunId;
-
-  if (!runId) {
-    throw new NotInWorkflowOrStepContextError(
-      'setAttributes()',
-      'https://workflow-sdk.dev/docs/api-reference/workflow/set-attributes'
-    );
-  }
-
-  await setAttributesStep(runId, changes);
+  throw new FatalError(
+    'setAttributes() can only be called from a step body in the V5 MVP. ' +
+      "Wrap it: `async function setAttrs(a) { 'use step'; await setAttributes(a); }` " +
+      'and call that from your workflow. The 5.0.0 attributes feature removes ' +
+      'this restriction; see the attributes-mvp changelog entry.'
+  );
 }
